+++ src/main/java/roomescape/controller/view/HomeViewController.java
@@ -1,18 +1,13 @@
-package roomescape.controller;
+package roomescape.controller.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller()
-public class PageController {
+public class HomeViewController {
 
     @GetMapping("/")
     public String readHomePage() {
         return "home";
-    }
-
-    @GetMapping("/reservation")
-    public String readReservationPage() {
-        return "reservation";
     }
 }
